+++ src/components/AdventureForm.jsx
@@ -1,16 +1,16 @@
-import { DateTimePickerHF } from "../../../components/hookForm/DateTimePickerHF";
-import { NumberInputHF } from "../../../components/hookForm/NumberInputHF";
-import { TextInputHF } from "../../../components/hookForm/TextInputHF";
-import { createAdventure } from "../../../services/app/adventures";
-import { LoadingDialog } from "../../../components/LoadingDialog";
-import { useErrorDialog } from "../../../hooks/useErrorDialog";
-import { ErrorDialog } from "../../../components/ErrorDialog";
+import { DateTimePickerHF } from "./hookForm/DateTimePickerHF";
+import { createAdventure } from "../services/app/adventures";
 import { Bonfire, Group, Map } from "iconoir-react-native";
+import { NumberInputHF } from "./hookForm/NumberInputHF";
+import { useErrorDialog } from "../hooks/useErrorDialog";
 import { Button, TextInput } from "react-native-paper";
-import { useLoading } from "../../../hooks/useLoading";
-import { sharedStyles } from "../../../shared/styles";
+import { TextInputHF } from "./hookForm/TextInputHF";
+import { useLoading } from "../hooks/useLoading";
+import { sharedStyles } from "../shared/styles";
+import { LoadingDialog } from "./LoadingDialog";
 import { useTranslation } from "react-i18next";
 import { Keyboard, View } from "react-native";
+import { ErrorDialog } from "./ErrorDialog";
 import { useForm } from "react-hook-form";
 
 function AdventureForm({ navigation }) {
@@ -96,7 +96,7 @@ function AdventureForm({ navigation }) {
         style={sharedStyles.mv15}
         uppercase={false}
         mode="contained">
-        {t("homeStackNavigator.createAdventure")}
+        {t("adventures.createAdventure")}
       </Button>
 
       <ErrorDialog isVisible={error} onDismiss={hideError} content={error} />