+++ src/screens/mainTabs/profile/Stack.jsx
@@ -1,5 +1,6 @@
 import { NavigationBar } from "../../../components/navigation/NavigationBar";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { AdventureForm } from "../../../components/AdventureForm";
 import { sharedStyles } from "../../../shared/styles";
 import { useTranslation } from "react-i18next";
 import { Edit } from "./Edit";
@@ -8,7 +9,7 @@ import { Profile } from ".";
 const Stack = createNativeStackNavigator();
 
 function ProfileStackNavigator() {
-  const { t } = useTranslation("translation", { keyPrefix: "components.profileStackNavigator" });
+  const { t } = useTranslation("translation", { keyPrefix: "components" });
 
   return (
     <Stack.Navigator
@@ -19,9 +20,17 @@ function ProfileStackNavigator() {
       initialRouteName="ProfileRoot">
       <Stack.Screen name="ProfileRoot" component={Profile} options={{ headerShown: false }} />
       <Stack.Screen
-        options={{ title: t("editProfile"), contentStyle: sharedStyles.ph20 }}
+        options={{ title: t("profileStackNavigator.editProfile"), contentStyle: sharedStyles.ph20 }}
         component={Edit}
         name="EditProfile"
+      />
+      <Stack.Screen
+        options={{
+          title: t("adventures.createAdventure"),
+          contentStyle: sharedStyles.ph20,
+        }}
+        component={AdventureForm}
+        name="AdventureForm"
       />
     </Stack.Navigator>
   );